pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sonos-lastfm"
version = "0.1.2"
description = "Automatically scrobble music playing on your Sonos speakers to Last.fm"
readme = "README.md"
authors = [
    { name = "Denis Moskalets", email = "denya.msk@gmail.com" }
]
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["sonos", "lastfm", "scrobbler", "music"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio",
]
dependencies = [
    "soco>=0.27.0",
    "pylast>=5.1.0",
    "python-dotenv>=1.0.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "keyring>=24.3.0",
]

[project.urls]
Homepage = "https://github.com/denya/sonos-lastfm"
Issues = "https://github.com/denya/sonos-lastfm/issues"

[project.scripts]
sonos-lastfm = "sonos_lastfm:main"

[tool.hatch.build]
include = [
    "src/sonos_lastfm/*.py",
    "LICENSE",
    "README.md"
]

[tool.hatch.build.targets.wheel]
packages = ["src/sonos_lastfm"]

[tool.mypy]
python_version = "0.1.2"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
show_error_codes = true

[tool.ruff]
target-version = "py312"
line-length = 88
lint.select = ["ALL"]
lint.ignore = ["PLW0603"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.pylint]
max-args = 6 